renaming pages for clarity and updating their imports as expected
