+++ src/main.tsx
@@ -8,7 +8,7 @@ import './index.css'
 import Index from './routes/Index';
 
 // Result Pages
-import ResultPage from './pages/ResultPage.tsx';
+import SingleResultPage from './pages/SingleResultPage.tsx';
 
 // Search Page Layout
 import SearchPageLayout from './pages/SearchPageLayout.tsx';
@@ -17,7 +17,7 @@ import SearchPageLayout from './pages/SearchPageLayout.tsx';
 import AnthologyPage from './pages/AnthologyPage.tsx';
 
 //Large Work Page
-import LargeWorkPage from './pages/LargeWork.tsx';
+import LargeWorkPage from './pages/LargeWorkPage.tsx';
 
 import { sermonRouteInfo, songRouteInfo, symphonyRouteInfo, analysisRouteInfo } from './routeInfo.ts'
 
@@ -31,16 +31,16 @@ createRoot(document.getElementById('root')!).render(
         <Route path="/" element={<App />}>
           <Route index element={<Index />} />
           <Route path="song" element={<SearchPageLayout routeInfo={songRouteInfo} />} />
-          <Route path="song/:songId" element={<ResultPage resultPageInfo={songResultInfo} />} />
+          <Route path="song/:songId" element={<SingleResultPage resultPageInfo={songResultInfo} />} />
           <Route path="song/:songId/anthology/:title" element={<AnthologyPage routeInfo={songRouteInfo} />} />
           <Route path="song/:songId/large-work/:title" element={<LargeWorkPage routeInfo={songRouteInfo} />} />
           <Route path="song-analysis" element={<SearchPageLayout routeInfo={analysisRouteInfo} />} />
-          <Route path="song-analysis/:songId" element={<ResultPage resultPageInfo={analysisResultInfo} />} />
+          <Route path="song-analysis/:songId" element={<SingleResultPage resultPageInfo={analysisResultInfo} />} />
           <Route path="song-analysis/:songId/anthology/:title" element={<AnthologyPage routeInfo={analysisRouteInfo} />} />
           <Route path="sermon" element={<SearchPageLayout routeInfo={sermonRouteInfo} />} />
-          <Route path="sermon/:sermonId" element={<ResultPage resultPageInfo={sermonResultInfo} />} />
+          <Route path="sermon/:sermonId" element={<SingleResultPage resultPageInfo={sermonResultInfo} />} />
           <Route path="symphony" element={<SearchPageLayout routeInfo={symphonyRouteInfo} />} />
-          <Route path="symphony/:symphonyId" element={<ResultPage resultPageInfo={symphonyResultInfo} />} />
+          <Route path="symphony/:symphonyId" element={<SingleResultPage resultPageInfo={symphonyResultInfo} />} />
         </Route>
       </Routes>
     </BrowserRouter>
