+++ src/pages/AnthologyPage.tsx
@@ -1,7 +1,7 @@
 import { useParams } from "react-router";
 import { useState, useEffect } from "react";
 import { searchSolr } from '../utils/utils';
-import SubPageResults from "./SubPageResults";
+import SubPageResults from "../components/SubPageResults";
 
 
 export default function AnthologyPage({ routeInfo }: any) {
